Plain HTML Version
This is the initial commit of the Backbone-Reactive demo. This consists
of two very simple html pages, one for the homepage/dashboard and one
for an inbox style list page.

The purpose of these pages are to serve as a last-resort fallback. If
the user, for whatever reason, doesn't have Javascript (possibly on an
old browser, old phone etc) then they can still see the basic details
needed.

With this demo project, I am not worried too much about css or even
valid/correct html. It is all about the Javascript, so please ignore the
lack of css and state of the html.

Included in this commit is a very simple static file server. This is to
allow for the easy testing of the demo, without having to mess about
with a "proper" web server, such as Apache or Nginx