Using a single Message View for all messages
My initial version for loading in the messages required creating
multiple instances of the MessageView.

This was unsatisfactory for two reasons:

1. This required some logic to determine if the MessageView was already
loaded on the page and if so replace the content, if not add the DOM
element.

2. Re-initialising the MessageView when viewing each message seems to be
a wast of effort. We only need one MessageView, so no need to have
multiple instance.

The way around this issue is to trigger an event when the user wishes to
view a message. This event should "bubble" up to top level the
ListView. The ListView has previously initiated a MessageView, so now it
is simply a matter of passing this existing MessageView the model that
the content is based on and then let Backbone re-render the Message View
as needed.